+++ LuckyoneUI/Core/Installer.lua
@@ -740,10 +740,9 @@ local function BuildPages()
 		-- Edit mode string and guide
 		Page(L["Blizzard Edit Mode"], {
 			Green(L["Step 1:"]) .. '\n' .. L["Click the first button for the import.\nUse CTRL+C to copy the string from the popup."],
-			Green(L["Step 2:"]) .. '\n' .. L["Enter Edit Mode and select Import on the Dropdown.\nUse CTRL+V to paste string, then pick a name and click import."],
+			Green(L["Step 2:"]) .. '\n' .. L["Press Escape, click Edit Mode and select Import on the Dropdown.\nUse CTRL+V to paste string, then pick a name and click import."],
 		}, {
 			Button(L["Copy Editmode String"], function() Private:Return_EditModeString() end, 'copy'),
-			Button(Green(L["Enter Edit Mode"]), function() Private:EnterEditMode() end),
 		}, nil, not Private.isRetail),
 
 		-- LuckyoneUI module checkboxes